+++ crates/sql/src/planner.rs
@@ -354,6 +354,7 @@ pub(crate) fn build_plan(
         PreparedKind::SetTransactionIsolation { .. } => {
             simple_node(PhysicalKind::Constant, "SET TRANSACTION".to_owned())
         }
+        PreparedKind::SetSessionVariable => simple_node(PhysicalKind::Constant, "SET".to_owned()),
         PreparedKind::ShowVariable { .. } => simple_node(PhysicalKind::Constant, "SHOW".to_owned()),
         PreparedKind::AlterIndex { .. } => {
             simple_node(PhysicalKind::Constant, "ALTER INDEX".to_owned())